Able to predict with tf lite model

+++ app/src/main/java/com/b21cap0398/acnescan/data/source/local/entity/Possibility.kt
@@ -2,5 +2,5 @@ package com.b21cap0398.acnescan.data.source.local.entity
 
 data class Possibility(
     val acne_name: String,
-    val possibility: Long
+    val possibility: Double
 )
